README.md
Capture the Flag (CTF) is an information security competition that challenges contestants to solve a variety of tasks. This project would be focused on creating a Jeopardy style CTF game, where individuals or groups of players would attempt to solve different scenarios, using Splunk, to earn points and compete against one another.